// src/types/brand.type.ts
export interface brandI {
    createdAt: string;
    image: string;
    name: string;
    slug: string;
    updatedAt: string;
    _id: string;
}
export interface AllBrandI {
    createdAt: string;
    image: string;
    name: string;
    slug: string;
    updatedAt: string;
    _id: string;
    imageCover: string;
    title: string;
    ratingsAverage: number;
    price: number;
}